pyproject.toml
[project]
name = "gui-grounding"
version = "0.1.0"
description = "Evaluate the performance of LLMs on GUI grounding tasks"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "openai>=1.63.1",
    "pillow>=11.1.0",
    "tqdm>=4.66.0",
    "boto3>=1.34.0",
]


[tool.ruff]
line-length = 100
